% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchSCore2Data-pkg.R
\docType{package}
\name{matchSCore2-pkg}
\alias{matchSCore2-pkg}
\title{matchSCore2Data: Accessing The Datasets Of matchSCore2 in R/Bioconductor}
\description{
\code{matchSCore2Data} is an ExperimentHub package which provides access to the single-cell
RNA-seq data from ...
}
\references{
If you use the data in this package, please refer to the original sources
as well (plus the related publications),
which are licensed under a Creative Commons Attribution 4.0 International License.
}
\author{
Federico Marini \email{marinif@uni-mainz.de}

Elisabetta Mereu \email{elisabetta.mereu@cnag.crg.eu}
}
