% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchSCore2Data.R
\name{matchSCore2Data}
\alias{matchSCore2Data}
\title{Download data from the matchSCore2 project via ExperimentHub}
\usage{
matchSCore2Data(dataset = NULL)
}
\arguments{
\item{dataset}{A character string: which dataset should be retrieved?}
}
\value{
A SingleCellExperiment object with ...
}
\description{
Download HDF5 (dense assay) and RDS (row and column annotations)
files from the HCA via ExperimentHub, composing them together as a
SingleCellExperiment object
}
\details{
This current release includes the following datasets:
}
\examples{
matchSCore2Data()
}
